--- pages/Create-Workout-Routine.js
@@ -1,13 +1,0 @@
-import Footer from '../components/Footer';
-import Header from '../components/Header';
-import CreateWorkoutRoutineForm from '../components/CreateGymPlan/CreateWorkoutRoutineForm/CreateWorkoutRoutineForm';
-
-export default function CreateGymPlan({ onCreatePlan }) {
-   return (
-      <>
-         <Header headline='create a workout routine' />
-         <CreateWorkoutRoutineForm onCreatePlan={onCreatePlan} />
-         <Footer />
-      </>
-   );
-}
